Change the return types for strtoq and strtouq to int64_t and u_int64_t instead of long long and unsigned long long. Really they should be quad_t and u_quad_t, but that would require sys/types.h and this header only includes machine/types.h. The difference here is that int64_t and u_int64_t on alpha are long and unsigned long, not long long etc. This is required to pass gcc's type checking where long != long long even though they are the same size of alpha.

+++ include/stdlib.h
@@ -169,9 +169,8 @@ void	 srandom __P((unsigned long));
 void	 srandomdev __P((void));
 char	*user_from_uid __P((unsigned long, int));
 #ifndef __STRICT_ANSI__
-long long
-	 strtoq __P((const char *, char **, int));
-unsigned long long
+int64_t	 strtoq __P((const char *, char **, int));
+u_int64_t
 	 strtouq __P((const char *, char **, int));
 #endif
 void	 unsetenv __P((const char *));
